requirements.txt
asn1crypto==0.24.0
astroid==2.1.0
auth0-python==3.6.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
clickclick==1.2.2
colorama==0.4.1
connexion==2.2.0
cryptography==2.5
ecdsa==0.13
Flask==1.0.2
future==0.17.1
idna==2.8
inflection==0.3.1
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
jose==1.0.0
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
numpy==1.16.1
openapi-spec-validator==0.2.4
pandas==0.24.1
pyasn1==0.4.5
pycparser==2.19
pylint==2.2.2
python-dateutil==2.8.0
python-jose==3.0.1
pytz==2018.9
PyYAML==3.13
requests==2.21.0
rsa==4.0
six==1.12.0
SQLAlchemy==1.2.17
swagger-ui-bundle==0.0.3
urllib3==1.24.1
Werkzeug==0.14.1
wincertstore==0.2
wrapt==1.11.0
